Refactored eigen_typekit and kdl_typekit to allow compilation in catkin devel-space
The main difference is that headers and IDL files are placed in the include/orocos folder of each package
so that other packages can use them directly without installation step.

Unfortunately the IDL compilers for the CORBA transport plugin do not support generation of header files in
a different output directory than source files. We need to copy them to their final destination after generation.

See also https://github.com/orocos/rtt_geometry/pull/19. This commit would be the alternative approach to
setting the declared build type to cmake in package.xml.
