+++ eigen_typekit/corba/CMakeLists.txt
@@ -8,9 +8,15 @@ if(OROCOS-RTT_CORBA_FOUND)
   set(ENABLE_CORBA TRUE)
   find_package(Corba REQUIRED)
 
-  FILE( GLOB IDLS [^.]*.idl )
+  FILE( GLOB IDLS
+    ${PROJECT_SOURCE_DIR}/include/orocos/${PROJECT_NAME}/transports/corba/[^.]*.idl
+  )
   FILE( GLOB CPPS [^.]*.cpp )
-  FILE( GLOB HPPS [^.]*.hpp [^.]*.h [^.]*.inl)
+  FILE( GLOB HPPS
+    ${PROJECT_SOURCE_DIR}/include/orocos/${PROJECT_NAME}/transports/corba/[^.]*.hpp
+    ${PROJECT_SOURCE_DIR}/include/orocos/${PROJECT_NAME}/transports/corba/[^.]*.h
+    ${PROJECT_SOURCE_DIR}/include/orocos/${PROJECT_NAME}/transports/corba/[^.]*.inl
+  )
 
   ORO_ADD_CORBA_SERVERS(CPPS HPPS ${IDLS} )
 
@@ -19,10 +25,9 @@ if(OROCOS-RTT_CORBA_FOUND)
   ADD_CUSTOM_TARGET( idl-eigen DEPENDS ${CPPS} ${HPPS} COMMENT "Generating sources from .idl files")
 
   INSTALL_FILES( /include/eigen/bindings/rtt/corba FILES ${HPPS})
-  INCLUDE_DIRECTORIES( ${CMAKE_CURRENT_BINARY_DIR} )
 
   orocos_use_package(orocos_eigen)
-  orocos_typekit(eigen-transport-corba SHARED ${CPPS})
+  orocos_typekit(eigen-transport-corba SHARED ${CPPS} ${HPPS})
   # Is used for building  the library
   TARGET_LINK_LIBRARIES(eigen-transport-corba ${OROCOS-RTT_CORBA_LIBRARIES})
