+++ eigen_typekit/corba/corbatoolkit.cpp
@@ -19,10 +19,10 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
-#include "corbatoolkit.hpp"
+#include <eigen_typekit/transports/corba/corbatoolkit.hpp>
 #include <rtt/types/Types.hpp>
 #include <rtt/transports/corba/CorbaTemplateProtocol.hpp>
-#include "CorbaEigenConversion.hpp"
+#include <eigen_typekit/transports/corba/CorbaEigenConversion.hpp>
 #include <rtt/types/TypekitPlugin.hpp>
 
 using namespace RTT;
