+++ eigen_typekit/eigen_typekit.cpp
@@ -17,7 +17,7 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
-#include "eigen_typekit.hpp"
+#include <eigen_typekit/eigen_typekit.hpp>
 #include <rtt/Property.hpp>
 #include <rtt/PropertyBag.hpp>
 #include <rtt/types/TemplateTypeInfo.hpp>
