+++ eigen_typekit/mqueue/CMakeLists.txt
@@ -8,7 +8,10 @@ if(OROCOS-RTT_MQUEUE_FOUND)
 
   target_link_libraries(eigen-transport-mqueue ${OROCOS-RTT_MQUEUE_LIBRARIES})
 
-  install(FILES eigen_mqueue.hpp DESTINATION include/orocos/eigen_typekit/transports/mqueue)
+  orocos_install_headers(
+    FILES ../include/orocos/${PROJECT_NAME}/transports/mqueue/eigen_mqueue.hpp
+    INSTALL include/orocos/${PROJECT_NAME}/transports/mqueue
+  )
 
 endif()
