+++ kdl_typekit/CMakeLists.txt
@@ -8,8 +8,12 @@ find_package(OROCOS-RTT REQUIRED)
 include(${OROCOS-RTT_USE_FILE_PATH}/UseOROCOS-RTT.cmake)
 
 find_package(orocos_kdl REQUIRED)
-include_directories(${orocos_kdl_INCLUDE_DIRS})
 link_directories(${orocos_kdl_LIBRARY_DIRS})
+
+include_directories(
+  include/orocos
+  ${orocos_kdl_INCLUDE_DIRS}
+)
 
 # Define the way rotation properties get written to XML
 # Set one of the two options below to write them as EULER ZYX or RPY
@@ -51,10 +55,9 @@ if(catkin_FOUND)
   set(LUA_DESTINATION ${CATKIN_GLOBAL_SHARE_DESTINATION})
 endif()
 
-install(
-  FILES src/kdlTypekitTypes.hpp
-  DESTINATION include/orocos/kdl_typekit/typekit
-  RENAME Types.hpp
+orocos_install_headers(
+  FILES include/orocos/${PROJECT_NAME}/typekit/Types.hpp
+  INSTALL include/orocos/${PROJECT_NAME}/typekit
 )
 install(
   FILES ${LUA_FILES}
