--- kdl_typekit/src/corba/orocos-kdltk-corba.pc.in
@@ -1,11 +1,0 @@
-prefix=@CMAKE_INSTALL_PREFIX@
-exec_prefix=${prefix}  # defining another variable in terms of the first
-libdir=${exec_prefix}/lib/rtt/@OROCOS_TARGET@/plugins
-includedir=${prefix}/include
-
-Name: orocos-kdltk-corba-@OROCOS_TARGET@ # human-readable name
-Description: The Orocos Kinematics and Dynamics Library RTT Plugin Corba  # human-readable description
-Requires: orocos-kdltk-@OROCOS_TARGET@ orocos-rtt-corba-@OROCOS_TARGET@
-Version: @KDL_VERSION@
-Libs: -L${libdir} -lorocos-kdltk-corba-@OROCOS_TARGET@
-Cflags:-I${includedir}
